Add Feature To Relate Weak Tables To Primary Ones
Added capability of retrieving audit information of weak entities'
tables querying straight from audit.logged_actions, when filtering by
primary entity's "code value" - a unique, human-attributed value
("code") that identifies such entity.

On table "logged_actions", the changes now show everything as
previously, with the addition of the key "Relationship" in column
"changed_fields", whose value indicates the primary table to which the
row of the weak one belongs.

It was done by storing information about audited tables of primary
entities and weak entities (in table "audit_metadata"), and the
relationship among them (in table "audit_reference").

The new function "audit_for_stronger_entity" recursively interacts with
the new tables as well as the audited ones in order to find out to
which primary table the row of the weak one belongs.

Since it is not always that a weak entity points to a stronger one,
"audit_reference" has a column to indicate the ones which are pointing
("direction"), and a column to indicate the name of the column that
references the pointed table ("relation_column").